backend/requirements.txt
fastapi==0.78.0
uvicorn==0.17.6
psycopg2-binary==2.9.3
gunicorn==20.1.0
wheel==0.37.1
requests==2.28.1
numpy==1.23.1
rasterio==1.3.0.post1
area==1.1.1